backend/requirements.txt
fastapi
uvicorn
PyYAML
pydantic
